Vector Plot outline
Have the basic vector plot outline code in place, just need to sort through and identify approach and departure cells from the p-values that are already loaded in. Also, there was a mistake when re-running the angle_distance_data so the partner, novel, and chamber data needs to be run again. Data given to Zoe was originally correct so no problem there.